requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Vector Database
qdrant-client==1.12.1

# AI & ML
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1
tf-keras==2.20.1
tensorflow==2.20.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
pandas==2.3.3
openpyxl==3.1.5

# HTTP Client
httpx==0.28.1

# AI Providers
mistralai>=0.0.7

# Cloud Storage
boto3>=1.34.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.4

# System Monitoring
psutil==7.0.0

# Development & Utilities
typing-extensions==4.15.0
numpy==2.3.4
pillow==11.3.0

# Optional: Enhanced Performance
# accelerate==1.2.1
# optimum==1.24.0

# Optional: Additional Document Formats
# python-pptx==1.0.2
# xlsxwriter==3.2.0
# beautifulsoup4==4.12.3
# lxml==6.0.2

# Advanced Features
spacy>=3.7.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
slowapi>=0.1.9
cryptography>=41.0.0
aiofiles>=23.2.1
matplotlib>=3.8.0
plotly>=5.17.0

# Optional: Enhanced Performance
# redis>=5.0.0
# accelerate==1.2.1
# optimum==1.24.0

# Optional: Monitoring & Logging
# prometheus-client==0.21.1
# structlog==24.4.0